check/072 is GFonts specific so move it away from OT spec.
(issue #1945)
